Moving DIR_Protocol to communicator.h
