+++ include/communicator.h
@@ -4,6 +4,7 @@
 #define __communicator_h
 
 #include <network.h>
+#include <machine/nic.h>
 #include <synchronizer.h>
 
 __BEGIN_SYS
@@ -270,6 +271,61 @@ public:
 
     Link<Channel> * listen() { return new (SYSTEM) Link<Channel>(Channel::listen(this->_local)); }
     Link<Channel> * connect(const Address & to) { return new (SYSTEM) Link<Channel>(Channel::connect(this->_local, to)); }
+};
+
+// A simple abstraction for connectionless channels, that simply forwards data
+// from the application to the next lower level. DIR_Protocol is responsible for
+// decoupling the application from the network communication.
+class DIR_Protocol:
+    private NIC<Ethernet>::Observer,
+    private Concurrent_Observer<Ethernet::Buffer, Ethernet::Protocol>
+{
+
+public:
+    typedef Ethernet::Protocol Protocol;
+    typedef Ethernet::Buffer Buffer;
+    typedef Ethernet::Address Address;
+    typedef Data_Observer<Buffer, Protocol> Observer;
+    typedef Data_Observed<Buffer, Protocol> Observed;
+
+public:
+    const unsigned int MTU        = Ethernet::MTU;
+    const unsigned short PROTOCOL = Ethernet::PROTO_DIR;
+
+    //template<unsigned int UNIT = 0>  see IP()
+    DIR_Protocol(unsigned int nic = 0) :
+            _nic(Traits<Ethernet>::DEVICES::Get<0>::Result::get(nic))
+    {
+        _nic->attach(this, PROTOCOL);
+    }
+
+    const Address & address() { return _nic->address(); }
+    const unsigned int mtu() { return this->MTU; }
+
+    int send(const void * data, unsigned int size) {
+        return _nic->send(_nic->broadcast(), PROTOCOL, data, size);
+    }
+
+    int receive(Address * src, void * data, unsigned int size) {
+        Buffer * buff = updated();
+        memcpy(data, buff->frame()->data<char>(), size);
+        _nic->free(buff);
+        return size;
+    }
+
+    void update(Observed* obs, const Protocol& prot, Buffer* buf) {
+        Concurrent_Observer<Observer::Observed_Data, Protocol>::update(prot, buf);
+    }
+
+    static bool notify(const Protocol & prot, Buffer * buf) {
+        return _observed.notify(prot, buf);
+    }
+
+protected:
+    NIC<Ethernet> * _nic;
+    Address _address;
+    static Observed _observed;
+
 };
 
 __END_SYS
